+++ pkgs/extension_discovery/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1-wip
 
 - Update the package description.
+- Support optional `packageUri`.
 
 ## 1.0.0